+++ CHANGELOG.md
@@ -11,6 +11,12 @@ future generated release notes can be inserted predictably.
 
 _No unreleased changes yet._
 
+## v0.29.0 (2026-06-27)
+
+### Features
+
+- add voice json output (3a5c35d)
+
 ## v0.28.0 (2026-06-27)
 
 ### Features